feat: add new example, rename todo example
